Refactor all examples into examples/
