+++ Cargo.toml
@@ -18,6 +18,7 @@ streaming-iterator = "0.1.5"
 bitflags = "1.2.1"
 
 [dev-dependencies]
+clap = "~2.27.0"
 serde = {version = "1.0.118", features = ["derive"]}
 bincode = "1.3.1"
 
@@ -25,3 +26,13 @@ bincode = "1.3.1"
 bindgen = "0.57.0"
 cc = { version = "1.0", features = ["parallel"] }
 pkg-config = "0.3"
+
+[[example]]
+name = "mutation_metadata_bincode"
+
+[[example]]
+name = "mutation_metadata_std"
+
+# Not run during tests
+[[example]]
+name = "tree_traversals"
