--- tskit_rust_examples/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "tskit_rust_examples"
-version = "0.1.0"
-authors = ["molpopgen <krthornt@uci.edu>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies.tskit]
-path = "../"
-
-[dependencies]
-clap = "~2.27.0"
-streaming-iterator = "0.1.5"
-
-[[bin]]
-name = "tree_traversals"
-path = "./tree_traversals.rs"
